modulo clasificacion completo con ajustes

+++ app/Responsable/Raees/RaeeDestroyResponsable.php
@@ -4,7 +4,7 @@ namespace App\Responsable\Raees;
 
 use Illuminate\Contracts\Support\Responsable;
 use App\Models\Raee;
-use Illuminate\Support\Facades\{DB};
+use Illuminate\Support\Facades\{Auth, DB};
 use App\Helpers\StandardResponse;
 use Illuminate\Http\Response;
 use App\Repositories\Raees\RaeeRepository;
@@ -23,7 +23,16 @@ class RaeeDestroyResponsable implements Responsable
     public function toResponse($request) {
         try {
             DB::beginTransaction();
-                $res = $this->repository->eliminarRaee($this->raee);
+                if (Auth::user()->enabled) {
+                    $res = $this->repository->eliminarRaee($this->raee);
+                }else{
+                    return response()->json([
+                        'success' => false,
+                        'code' =>  Response::HTTP_UNAUTHORIZED,
+                        'message' => 'No estás habilitado para esta acción.',
+                        'data' => []
+                    ],Response::HTTP_UNAUTHORIZED);
+                }
             DB::commit();
 			return $this->destroyResponse($res ? Response::HTTP_OK : Response::HTTP_INTERNAL_SERVER_ERROR);
         } catch (\Throwable $e) {
